Move format related stuff from chunk
